src/backend/requirements.txt
Django==1.11.5
channels==1.1.7
asgi_redis==1.4.2
pytest-django==3.1.2
factory_boy==2.9.2
